Extracted sendEmail and slightly updated email copy
