+++ src/services/sendEmail.js
@@ -1,0 +1,18 @@
+// using Twilio SendGrid's v3 Node.js Library
+// https://github.com/sendgrid/sendgrid-nodejs
+const sgMail = require('@sendgrid/mail');
+
+sgMail.setApiKey(process.env.SENDGRID_API_KEY);
+
+const sendEmail = msg => {
+  sgMail
+    .send(msg)
+    .then(() => {
+      console.log(`Email sent to ${msg.to}`);
+    })
+    .catch(error => {
+      console.error(error);
+    });
+};
+
+exports.sendEmail = sendEmail;
